+++ connect/src/spirc.rs
@@ -4,7 +4,7 @@ use crate::{
         authentication::Credentials,
         dealer::{
             manager::{BoxedStream, BoxedStreamResult, Reply, RequestReply},
-            protocol::{Command, Message, Request},
+            protocol::{Command, FallbackWrapper, Message, Request},
         },
         session::UserAttributes,
         Error, Session, SpotifyId,
@@ -28,9 +28,10 @@ use crate::{
         provider::IsProvider,
         {ConnectConfig, ConnectState},
     },
-    LoadContextOptions, LoadRequestOptions,
+    LoadContextOptions, LoadRequestOptions, PlayContext,
 };
 use futures_util::StreamExt;
+use librespot_protocol::context_page::ContextPage;
 use protobuf::MessageField;
 use std::{
     future::Future,
@@ -81,7 +82,7 @@ struct SpircTask {
     connect_state_volume_update: BoxedStreamResult<SetVolumeCommand>,
     connect_state_logout_request: BoxedStreamResult<LogoutCommand>,
     playlist_update: BoxedStreamResult<PlaylistModificationInfo>,
-    session_update: BoxedStreamResult<SessionUpdate>,
+    session_update: BoxedStreamResult<FallbackWrapper<SessionUpdate>>,
     connect_state_command: BoxedStream<RequestReply>,
     user_attributes_update: BoxedStreamResult<UserAttributesUpdate>,
     user_attributes_mutation: BoxedStreamResult<UserAttributesMutation>,
@@ -153,7 +154,7 @@ impl TryFrom<&str> for SpircCommand {
 const CONTEXT_FETCH_THRESHOLD: usize = 2;
 
 // delay to update volume after a certain amount of time, instead on each update request
-const VOLUME_UPDATE_DELAY: Duration = Duration::from_secs(2);
+const VOLUME_UPDATE_DELAY: Duration = Duration::from_millis(500);
 // to reduce updates to remote, we group some request by waiting for a set amount of time
 const UPDATE_STATE_DELAY: Duration = Duration::from_millis(200);
 
@@ -211,7 +212,7 @@ impl Spirc {
 
         let session_update = session
             .dealer()
-            .listen_for("social-connect/v2/session_update", Message::from_json)?;
+            .listen_for("social-connect/v2/session_update", Message::try_from_json)?;
 
         let user_attributes_update = session
             .dealer()
@@ -665,9 +666,9 @@ impl SpircTask {
             SpircCommand::Next => self.handle_next(None)?,
             SpircCommand::VolumeUp => self.handle_volume_up(),
             SpircCommand::VolumeDown => self.handle_volume_down(),
-            SpircCommand::Shuffle(shuffle) => self.connect_state.handle_shuffle(shuffle)?,
-            SpircCommand::Repeat(repeat) => self.connect_state.set_repeat_context(repeat),
-            SpircCommand::RepeatTrack(repeat) => self.connect_state.set_repeat_track(repeat),
+            SpircCommand::Shuffle(shuffle) => self.handle_shuffle(shuffle)?,
+            SpircCommand::Repeat(repeat) => self.handle_repeat_context(repeat)?,
+            SpircCommand::RepeatTrack(repeat) => self.handle_repeat_track(repeat),
             SpircCommand::SetPosition(position) => self.handle_seek(position),
             SpircCommand::SetVolume(volume) => self.set_volume(volume),
             SpircCommand::Load(command) => self.handle_load(command, None).await?,
@@ -999,12 +1000,24 @@ impl SpircTask {
                 self.handle_transfer(transfer.data.expect("by condition checked"))?;
                 return self.notify().await;
             }
-            Play(play) => {
-                let context_uri = play
-                    .context
-                    .uri
-                    .clone()
-                    .ok_or(SpircError::NoUri("context"))?;
+            Play(mut play) => {
+                if !self.connect_state.is_active() {
+                    self.handle_activate()
+                }
+
+                let context = match play.context.uri {
+                    Some(s) => PlayContext::Uri(s),
+                    None if !play.context.pages.is_empty() => PlayContext::Tracks(
+                        play.context
+                            .pages
+                            .iter()
+                            .cloned()
+                            .flat_map(|p| p.tracks)
+                            .flat_map(|t| t.uri)
+                            .collect(),
+                    ),
+                    None => Err(SpircError::NoUri("context"))?,
+                };
 
                 let context_options = play
                     .options
@@ -1013,16 +1026,16 @@ impl SpircTask {
                     .map(LoadContextOptions::Options);
 
                 self.handle_load(
-                    LoadRequest::from_context_uri(
-                        context_uri,
-                        LoadRequestOptions {
+                    LoadRequest {
+                        context,
+                        options: LoadRequestOptions {
                             start_playing: true,
                             seek_to: play.options.seek_to.unwrap_or_default(),
                             playing_track: play.options.skip_to.and_then(|s| s.try_into().ok()),
                             context_options,
                         },
-                    ),
-                    Some(play.context),
+                    },
+                    play.context.pages.pop(),
                 )
                 .await?;
 
@@ -1034,23 +1047,25 @@ impl SpircTask {
                 trace!("seek to {seek_to:?}");
                 self.handle_seek(seek_to.value)
             }
-            SetShufflingContext(shuffle) => self.connect_state.handle_shuffle(shuffle.value)?,
-            SetRepeatingContext(repeat_context) => self
-                .connect_state
-                .handle_set_repeat(Some(repeat_context.value), None)?,
-            SetRepeatingTrack(repeat_track) => self
-                .connect_state
-                .handle_set_repeat(None, Some(repeat_track.value))?,
+            SetShufflingContext(shuffle) => self.handle_shuffle(shuffle.value)?,
+            SetRepeatingContext(repeat_context) => {
+                self.handle_repeat_context(repeat_context.value)?
+            }
+            SetRepeatingTrack(repeat_track) => self.handle_repeat_track(repeat_track.value),
             AddToQueue(add_to_queue) => self.connect_state.add_to_queue(add_to_queue.track, true),
             SetQueue(set_queue) => self.connect_state.handle_set_queue(set_queue),
             SetOptions(set_options) => {
-                let context = set_options.repeating_context;
-                let track = set_options.repeating_track;
-                self.connect_state.handle_set_repeat(context, track)?;
+                if let Some(repeat_context) = set_options.repeating_context {
+                    self.handle_repeat_context(repeat_context)?
+                }
+
+                if let Some(repeat_track) = set_options.repeating_track {
+                    self.handle_repeat_track(repeat_track)
+                }
 
                 let shuffle = set_options.shuffling_context;
                 if let Some(shuffle) = shuffle {
-                    self.connect_state.handle_shuffle(shuffle)?;
+                    self.handle_shuffle(shuffle)?;
                 }
             }
             SkipNext(skip_next) => self.handle_next(skip_next.track.map(|t| t.uri))?,
@@ -1068,6 +1083,8 @@ impl SpircTask {
     fn handle_transfer(&mut self, mut transfer: TransferState) -> Result<(), Error> {
         let mut ctx_uri = match transfer.current_session.context.uri {
             None => Err(SpircError::NoUri("transfer context"))?,
+            // can apparently happen when a state is transferred stared with "uris" via the api
+            Some(ref uri) if uri == "-" => String::new(),
             Some(ref uri) => uri.clone(),
         };
 
@@ -1088,6 +1105,27 @@ impl SpircTask {
         }
 
         let fallback = self.connect_state.current_track(|t| &t.uri).clone();
+        let load_from_context_uri = !ctx_uri.is_empty();
+
+        if load_from_context_uri {
+            self.context_resolver.add(ResolveContext::from_uri(
+                ctx_uri.clone(),
+                &fallback,
+                ContextType::Default,
+                ContextAction::Replace,
+            ));
+        } else {
+            self.load_context_from_tracks(
+                transfer
+                    .current_session
+                    .context
+                    .pages
+                    .iter()
+                    .cloned()
+                    .flat_map(|p| p.tracks)
+                    .collect::<Vec<_>>(),
+            )?
+        }
 
         self.context_resolver.add(ResolveContext::from_uri(
             ctx_uri.clone(),
@@ -1134,7 +1172,15 @@ impl SpircTask {
             ))
         }
 
-        self.transfer_state = Some(transfer);
+        if load_from_context_uri {
+            self.transfer_state = Some(transfer);
+        } else {
+            let ctx = self.connect_state.get_context(ContextType::Default)?;
+            let idx = ConnectState::find_index_in_context(ctx, |pt| {
+                self.connect_state.current_track(|t| pt.uri == t.uri)
+            })?;
+            self.connect_state.reset_playback_to_position(Some(idx))?;
+        }
 
         self.load_track(is_playing, position.try_into()?)
     }
@@ -1203,61 +1249,41 @@ impl SpircTask {
     async fn handle_load(
         &mut self,
         cmd: LoadRequest,
-        context: Option<Context>,
+        page: Option<ContextPage>,
     ) -> Result<(), Error> {
         self.connect_state
-            .reset_context(ResetContext::WhenDifferent(&cmd.context_uri));
+            .reset_context(if let PlayContext::Uri(ref uri) = cmd.context {
+                ResetContext::WhenDifferent(uri)
+            } else {
+                ResetContext::Completely
+            });
 
         self.connect_state.reset_options();
 
-        if !self.connect_state.is_active() {
-            self.handle_activate();
-        }
-
-        let fallback = if let Some(ref ctx) = context {
-            match ConnectState::get_context_uri_from_context(ctx) {
-                Some(ctx_uri) => ctx_uri,
-                None => Err(SpircError::InvalidUri(cmd.context_uri.clone()))?,
+        let autoplay = matches!(cmd.context_options, Some(LoadContextOptions::Autoplay));
+        match cmd.context {
+            PlayContext::Uri(uri) => {
+                self.load_context_from_uri(uri, page.as_ref(), autoplay)
+                    .await?
             }
-        } else {
-            &cmd.context_uri
-        };
-
-        let update_context = if matches!(cmd.context_options, Some(LoadContextOptions::Autoplay)) {
-            ContextType::Autoplay
-        } else {
-            ContextType::Default
-        };
-
-        self.connect_state.set_active_context(update_context);
-
-        let current_context_uri = self.connect_state.context_uri();
-        if current_context_uri == &cmd.context_uri && fallback == cmd.context_uri {
-            debug!("context <{current_context_uri}> didn't change, no resolving required")
-        } else {
-            debug!("resolving context for load command");
-            self.context_resolver.clear();
-            self.context_resolver.add(ResolveContext::from_uri(
-                &cmd.context_uri,
-                fallback,
-                update_context,
-                ContextAction::Replace,
-            ));
-            let context = self.context_resolver.get_next_context(Vec::new).await;
-            self.handle_next_context(context);
+            PlayContext::Tracks(tracks) => self.load_context_from_tracks(tracks)?,
         }
+
+        let cmd_options = cmd.options;
+
+        self.connect_state.set_active_context(ContextType::Default);
 
         // for play commands with skip by uid, the context of the command contains
         // tracks with uri and uid, so we merge the new context with the resolved/existing context
-        self.connect_state.merge_context(context);
+        self.connect_state.merge_context(page);
 
         // load here, so that we clear the queue only after we definitely retrieved a new context
         self.connect_state.clear_next_tracks();
         self.connect_state.clear_restrictions();
 
-        debug!("play track <{:?}>", cmd.playing_track);
+        debug!("play track <{:?}>", cmd_options.playing_track);
 
-        let index = match cmd.playing_track {
+        let index = match cmd_options.playing_track {
             None => None,
             Some(ref playing_track) => Some(match playing_track {
                 PlayingTrack::Index(i) => *i as usize,
@@ -1272,7 +1298,7 @@ impl SpircTask {
             }),
         };
 
-        if let Some(LoadContextOptions::Options(ref options)) = cmd.context_options {
+        if let Some(LoadContextOptions::Options(ref options)) = cmd_options.context_options {
             debug!(
                 "loading with shuffle: <{}>, repeat track: <{}> context: <{}>",
                 options.shuffle, options.repeat, options.repeat_track
@@ -1283,7 +1309,8 @@ impl SpircTask {
             self.connect_state.set_repeat_track(options.repeat_track);
         }
 
-        if matches!(cmd.context_options, Some(LoadContextOptions::Options(ref o)) if o.shuffle) {
+        if matches!(cmd_options.context_options, Some(LoadContextOptions::Options(ref o)) if o.shuffle)
+        {
             if let Some(index) = index {
                 self.connect_state.set_current_track(index)?;
             } else {
@@ -1304,11 +1331,72 @@ impl SpircTask {
         }
 
         if self.connect_state.current_track(MessageField::is_some) {
-            self.load_track(cmd.start_playing, cmd.seek_to)?;
+            self.load_track(cmd_options.start_playing, cmd_options.seek_to)?;
         } else {
             info!("No active track, stopping");
             self.handle_stop()
         }
+
+        Ok(())
+    }
+
+    async fn load_context_from_uri(
+        &mut self,
+        context_uri: String,
+        page: Option<&ContextPage>,
+        autoplay: bool,
+    ) -> Result<(), Error> {
+        if !self.connect_state.is_active() {
+            self.handle_activate();
+        }
+
+        let update_context = if autoplay {
+            ContextType::Autoplay
+        } else {
+            ContextType::Default
+        };
+
+        self.connect_state.set_active_context(update_context);
+
+        let fallback = match page {
+            // check that the uri is valid or the page has a valid uri that can be used
+            Some(page) => match ConnectState::find_valid_uri(Some(&context_uri), Some(page)) {
+                Some(ctx_uri) => ctx_uri,
+                None => return Err(SpircError::InvalidUri(context_uri).into()),
+            },
+            // when there is no page, the uri should be valid
+            None => &context_uri,
+        };
+
+        let current_context_uri = self.connect_state.context_uri();
+
+        if current_context_uri == &context_uri && fallback == context_uri {
+            debug!("context <{current_context_uri}> didn't change, no resolving required")
+        } else {
+            debug!("resolving context for load command");
+            self.context_resolver.clear();
+            self.context_resolver.add(ResolveContext::from_uri(
+                &context_uri,
+                fallback,
+                update_context,
+                ContextAction::Replace,
+            ));
+            let context = self.context_resolver.get_next_context(Vec::new).await;
+            self.handle_next_context(context);
+        }
+
+        Ok(())
+    }
+
+    fn load_context_from_tracks(&mut self, tracks: impl Into<ContextPage>) -> Result<(), Error> {
+        let ctx = Context {
+            pages: vec![tracks.into()],
+            ..Default::default()
+        };
+
+        let _ = self
+            .connect_state
+            .update_context(ctx, ContextType::Default)?;
 
         Ok(())
     }
@@ -1405,6 +1493,23 @@ impl SpircTask {
         };
     }
 
+    fn handle_shuffle(&mut self, shuffle: bool) -> Result<(), Error> {
+        self.player.emit_shuffle_changed_event(shuffle);
+        self.connect_state.handle_shuffle(shuffle)
+    }
+
+    fn handle_repeat_context(&mut self, repeat: bool) -> Result<(), Error> {
+        self.player
+            .emit_repeat_changed_event(repeat, self.connect_state.repeat_track());
+        self.connect_state.handle_set_repeat_context(repeat)
+    }
+
+    fn handle_repeat_track(&mut self, repeat: bool) {
+        self.player
+            .emit_repeat_changed_event(self.connect_state.repeat_context(), repeat);
+        self.connect_state.set_repeat_track(repeat);
+    }
+
     fn handle_preload_next_track(&mut self) {
         // Requests the player thread to preload the next track
         match self.play_status {
@@ -1438,7 +1543,8 @@ impl SpircTask {
         let require_load_new = !self
             .connect_state
             .has_next_tracks(Some(CONTEXT_FETCH_THRESHOLD))
-            && self.session.autoplay();
+            && self.session.autoplay()
+            && !self.connect_state.context_uri().is_empty();
 
         if !require_load_new {
             return;
@@ -1519,16 +1625,16 @@ impl SpircTask {
     }
 
     fn handle_volume_up(&mut self) {
-        let volume_steps = self.connect_state.device_info().capabilities.volume_steps as u16;
+        let volume = (self.connect_state.device_info().volume as u16)
+            .saturating_add(self.connect_state.volume_step_size);
 
-        let volume = (self.connect_state.device_info().volume as u16).saturating_add(volume_steps);
         self.set_volume(volume);
     }
 
     fn handle_volume_down(&mut self) {
-        let volume_steps = self.connect_state.device_info().capabilities.volume_steps as u16;
+        let volume = (self.connect_state.device_info().volume as u16)
+            .saturating_sub(self.connect_state.volume_step_size);
 
-        let volume = (self.connect_state.device_info().volume as u16).saturating_sub(volume_steps);
         self.set_volume(volume);
     }
 
@@ -1557,7 +1663,24 @@ impl SpircTask {
         Ok(())
     }
 
-    fn handle_session_update(&mut self, mut session_update: SessionUpdate) {
+    fn handle_session_update(&mut self, session_update: FallbackWrapper<SessionUpdate>) {
+        // we know that this enum value isn't present in our current proto definitions, by that
+        // the json parsing fails because the enum isn't known as proto representation
+        const WBC: &str = "WIFI_BROADCAST_CHANGED";
+
+        let mut session_update = match session_update {
+            FallbackWrapper::Inner(update) => update,
+            FallbackWrapper::Fallback(value) => {
+                let fallback_inner = value.to_string();
+                if fallback_inner.contains(WBC) {
+                    log::debug!("Received SessionUpdate::{WBC}");
+                } else {
+                    log::warn!("SessionUpdate couldn't be parse correctly: {value:?}");
+                }
+                return;
+            }
+        };
+
         let reason = session_update.reason.enum_value();
 
         let mut session = match session_update.session.take() {
@@ -1644,6 +1767,8 @@ impl SpircTask {
     }
 
     fn set_volume(&mut self, volume: u16) {
+        debug!("SpircTask::set_volume({})", volume);
+
         let old_volume = self.connect_state.device_info().volume;
         let new_volume = volume as u32;
         if old_volume != new_volume || self.mixer.volume() != volume {