+++ playback/src/player.rs
@@ -85,6 +85,7 @@ struct PlayerInternal {
 
     player_id: usize,
     play_request_id_generator: SeqGenerator<u64>,
+    last_progress_update: Instant,
 }
 
 static PLAYER_COUNTER: AtomicUsize = AtomicUsize::new(0);
@@ -191,6 +192,14 @@ pub enum PlayerEvent {
         volume: u16,
     },
     PositionCorrection {
+        play_request_id: u64,
+        track_id: SpotifyId,
+        position_ms: u32,
+    },
+    /// Requires `PlayerConfig::position_update_interval` to be set to Some.
+    /// Once set this event will be sent periodically while playing the track to inform about the
+    /// current playback position
+    PositionChanged {
         play_request_id: u64,
         track_id: SpotifyId,
         position_ms: u32,
@@ -600,6 +609,7 @@ impl Player {
 
                 player_id,
                 play_request_id_generator: SeqGenerator::new(0),
+                last_progress_update: Instant::now(),
             };
 
             // While PlayerInternal is written as a future, it still contains blocking code.
@@ -1036,6 +1046,10 @@ impl PlayerTrackLoader {
             AudioFileFormat::MP3_160_ENC => 20.,
             AudioFileFormat::AAC_24 => 3.,
             AudioFileFormat::AAC_48 => 6.,
+            AudioFileFormat::AAC_160 => 20.,
+            AudioFileFormat::AAC_320 => 40.,
+            AudioFileFormat::MP4_128 => 16.,
+            AudioFileFormat::OTHER5 => 40.,
             AudioFileFormat::FLAC_FLAC => 112., // assume 900 kbit/s on average
             AudioFileFormat::XHE_AAC_12 => 1.5,
             AudioFileFormat::XHE_AAC_16 => 2.,
@@ -1454,6 +1468,22 @@ impl Future for PlayerInternal {
                                                     track_id,
                                                     position_ms: new_stream_position_ms,
                                                 });
+                                            }
+
+                                            if let Some(interval) =
+                                                self.config.position_update_interval
+                                            {
+                                                let last_progress_update_since_ms =
+                                                    now.duration_since(self.last_progress_update);
+
+                                                if last_progress_update_since_ms > interval {
+                                                    self.last_progress_update = now;
+                                                    self.send_event(PlayerEvent::PositionChanged {
+                                                        play_request_id,
+                                                        track_id,
+                                                        position_ms: new_stream_position_ms,
+                                                    });
+                                                }
                                             }
                                         }
                                         Err(e) => {
@@ -1899,7 +1929,7 @@ impl PlayerInternal {
             self.ensure_sink_stopped(play);
         }
 
-        if matches!(self.state, PlayerState::Invalid { .. }) {
+        if matches!(self.state, PlayerState::Invalid) {
             return Err(Error::internal(format!(
                 "Player::handle_command_load called from invalid state: {:?}",
                 self.state